test(runtime): isolate critical-section regressions
